COMMIT TO REVIEW many uncommited changes from IEEE challenge trials

+++ Classification/include/CGAL/Classification/K_means_classifier.h
@@ -82,7 +82,7 @@ public:
 
   /// \cond SKIP_IN_MANUAL
   void operator() (std::size_t item_index,
-                   std::vector<float>& out) const
+                   std::vector<double>& out) const
   {
     out.resize (m_labels.size());
     float total = 0.f;
